refactor(Slider): 슬라이더 컴포넌트 재구성 및 스타일 제거
